requirements.txt
fastapi
uvicorn
python-dotenv
praw
numpy
scikit-learn
sentence-transformers
supabase
google-generativeai
typing-extensions